Work in progress on moving new felix code ito sensible modules.
