+++ setup.py
@@ -71,7 +71,6 @@ setuptools.setup(
         'console_scripts': [
             'calico-acl-manager = calico.acl_manager.acl_manager:main',
             'calico-felix = calico.felix.felix:main',
-            'calico-felix2 = calico.felix.felix2:main',
         ],
         'neutron.ml2.mechanism_drivers': [
             'calico = calico.openstack.mech_calico:CalicoMechanismDriver',
